readme
Used Retrofit library for api calling

Retrofit is a REST Client for Android and Java by Square. It makes it relatively easy to retrieve and upload JSON (or other structured data) via a REST based webservice. Retrofit uses the OkHttp library for HTTP requests. Retrofit can be configured with converter is used for its data serialization.

# Glide is used for image processing